+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the ability to load official TinyMCE translation files from the `config/tinymce/languages` directory
 - Added `spicyweb\tinymce\services\LanguageService`
 
+### Changed
+- Updated the distributed TinyMCE to 6.5.1
+
 ## 1.3.0 - 2023-02-15
 
 ### Added